Project 5: Groundhog Calendar
Groundhog Calendar (Groundhog.py): A groundhog calendar is one where the amount of days in the month is always divisible by 7 (7,14,21,28). This program gets the number of months on the groundhog calendar and days in the month (You are free to enter a number that is not divisible by 7) and it will return a grid with each month and its corresponding days.