Project 6: Number Guessing Game
Number Guessing Game (numguessgame.py): This program picks a random number from 1 to 100 and the user must guess what the number is. If there are more than 3 guesses made that are greater than the computer-generated value, the user loses. There can be as many low guesses as the user wants, only high guesses are noted.