Project 8: Closing Time
Closing Time(closing.py): This program was made to figure out how long people who are working at a voting station during the election would have to stay. There are only a certain number of booths for many people to use and these people make sure to take their time when voting. Given how long each person takes to vote and the number of booths, this program can generate the amount of time the employees will have to stay.