Project 10: Covid Results List
Covid Results List(report.py): This program reads in a file that has a list of names and results from a covid test. It will add the names of people and their results if they are not already on the list. If someone was already on the list, their results will only change if the test came back as positive. It will return all of this in an output file.